Feature/all uses approuter (#217)
* Use const instead of let

* Change to app router
