+++ app/settings/page.tsx
@@ -11,8 +11,8 @@ export const metadata: Metadata = {
 };
 async function Settings({}) {
   const session = await getServerSession(authOptions);
-  let userEmail = session?.user?.email;
-  let userName = session?.user?.name;
+  const userEmail = session?.user?.email;
+  const userName = session?.user?.name;
   // if not logged in, do not show anything
   if (!session) return <LogInBtn />;
